Avances de la sesión del 9 de noviembre
